+++ src/components/Forms/AuthorizationForm/LogIn/LogIn.tsx
@@ -3,29 +3,18 @@ import React from 'react';
 import {Button, IconButton, InputAdornment, OutlinedInput, TextField} from "@mui/material";
 import Visibility from '@mui/icons-material/Visibility';
 import VisibilityOff from '@mui/icons-material/VisibilityOff';
-import {useRouter} from "../../../../hooks/Router/Router";
 import {LOG_IN} from "../../../../constants/Constants";
 import Title from "../Title/Title";
+import useLogIn from "../../../../hooks/LogIn/LogIn";
 
 export default function LogIn() {
 
-    // Todo: create separate hook when refactoring
-    const router = useRouter();
-    const [showPassword, setShowPassword] = React.useState<boolean>(false);
-
-    const handleClickShowPassword = () => {
-        setShowPassword((prev) => (!prev));
-    };
-
-    const handleMouseDownPassword = (event: React.MouseEvent<HTMLButtonElement>) => {
-        event.preventDefault();
-    };
-
-    const handleSubmit = (e: { preventDefault: () => void; }) => {
-        console.log('This thing was submitted!');
-        router.push('/main');
-        e.preventDefault();
-    };
+    const {
+        showPassword,
+        handleClickShowPassword,
+        handleMouseDownPassword,
+        handleSubmit
+    } = useLogIn();
 
     return (
         <>
