+++ src/constants/Constants.ts
@@ -1,10 +1,9 @@
 export const PAGES = ['Practice', 'Testing', 'Blog', 'Contact', 'Support'];
 export const SETTINGS = ['Profile', 'Progress', 'Account', 'Logout'];
-// Todo: rename or delete this variables
-export const HEADER_LOGO = "GOXTYPE";
+export const SITE_NAME = "GOXTYPE";
+
 export const AUTH = 0;
 export const REG = 1;
 
-
 export const LOG_IN = "Log In";
-export const SIGN_UP = "Sing Up";
+export const SIGN_UP = "Sign Up";
